+++ FishingAssistant/Debugging/DebugWarpPolicy.cs
@@ -1,9 +1,0 @@
-namespace FishingAssistant.Debugging;
-
-internal static class DebugWarpPolicy
-{
-    public static bool CanWarp(bool isWorldReady, bool hasActiveEvent, bool hasActiveMinigame)
-    {
-        return isWorldReady && !hasActiveEvent && !hasActiveMinigame;
-    }
-}